ジョブクローラ 保存済ページID取得

+++ batch/goBatch/internal/crawler/repository/repository_core.go
@@ -8,7 +8,7 @@ type GuitarRepository interface {
 
 type JobRepository interface {
 	Saver[*model.Job]
-	Selector[string, map[int64]struct{}]
+	Selector[string, map[int]struct{}]
 }
 
 type Saver[T any] interface {
